Reversing Kevin's checkin with regards to:
#4588 - Uninstaller is not deleting Martus group in Start menu in Windows 7 and Vista
#5138 - New NSIS uninstall wasn't always removing start menu group

With windows 7, we already must have admin privileges, so this change is fine, and does fix this issue only if the install put the files there to begin with.  So a previous install of Martus must manually remove those old links and folders from the start menu and then from now one the un-installer will remove those links etc from the start menu with this code change.